requirements.txt
fastapi==0.115.12
uvicorn==0.32.1
python_multipart==0.0.20
numpy==1.25.0
pillow==11.1.0
onnxruntime==1.19.0
boto3==1.37.32